requirements.txt
Filetype==1.0.7
numpy>=2
opencv-python==4.12.0.88
pandas==2.3.1
Pillow==11.3.0
PyMuPDF==1.26.3 
pytesseract==0.3.13
